refactor(example): code review cleanup
