+++ examples/oci-registry/src/storage.rs
@@ -1,6 +1,6 @@
-//! The storage layer over async `wasmcloud:blobstore`, plus the object-key naming
-//! scheme. Everything the registry persists lives as an object in a single
-//! container, keyed by repository name.
+//! The storage layer over async `wasmcloud:blobstore`. Everything the registry
+//! persists lives as an object in a single container, keyed by repository name
+//! (see [`crate::keys`] for the naming scheme).
 
 use crate::Container;
 use crate::bindings;
@@ -153,44 +153,4 @@ pub(crate) fn blob_err(error: BlobError) -> String {
         BlobError::QuotaExceeded => "blobstore quota exceeded".to_string(),
         BlobError::Other(message) => message,
     }
-}
-
-// --- key naming --------------------------------------------------------------
-
-/// Digests carry a `:` separator (`sha256:<hex>`) which is not portable across
-/// blobstore backends (e.g. NATS object store keys), so it is sanitized.
-fn sanitize(reference: &str) -> String {
-    reference.replace(':', "_")
-}
-
-pub(crate) fn blob_key(name: &str, digest: &str) -> String {
-    format!("{name}/blobs/{}", sanitize(digest))
-}
-
-pub(crate) fn manifest_key(name: &str, digest: &str) -> String {
-    format!("{name}/manifests/{}", sanitize(digest))
-}
-
-pub(crate) fn media_type_key(manifest_key: &str) -> String {
-    format!("{manifest_key}.mediatype")
-}
-
-pub(crate) fn tag_key(name: &str, tag: &str) -> String {
-    format!("{name}/tags/{tag}")
-}
-
-pub(crate) fn upload_key(name: &str, session: &str) -> String {
-    format!("{name}/uploads/{session}")
-}
-
-pub(crate) fn referrer_prefix(name: &str, subject: &str) -> String {
-    format!("{name}/referrers/{}", sanitize(subject))
-}
-
-pub(crate) fn referrer_key(name: &str, subject: &str, manifest_digest: &str) -> String {
-    format!(
-        "{}/{}",
-        referrer_prefix(name, subject),
-        sanitize(manifest_digest)
-    )
 }
