+++ examples/oci-registry/wkg.lock
@@ -1,0 +1,30 @@
+# This file is automatically generated.
+# It is not intended for manual editing.
+version = 1
+
+[[packages]]
+name = "wasi:http"
+registry = "wasi.dev"
+
+[[packages.versions]]
+requirement = "=0.3.0"
+version = "0.3.0"
+digest = "sha256:92cd8f3730c00226dc15626a2e7b21834dd187fc221f09818720d228585bbbf7"
+
+[[packages]]
+name = "wasi:random"
+registry = "wasi.dev"
+
+[[packages.versions]]
+requirement = "=0.3.0"
+version = "0.3.0"
+digest = "sha256:daa3c8897a68fbfc58b95c2e1de1b7af5be8c647bbb2cfeaa54c1e5b1ba9b772"
+
+[[packages]]
+name = "wasmcloud:blobstore"
+registry = "wasmcloud.com"
+
+[[packages.versions]]
+requirement = "=0.1.0"
+version = "0.1.0"
+digest = "sha256:c50a61278a9bfa39de81ebe2bcd8ca654ba96767624454d12d9e1fe62436a7b0"
